feat(mcp): support STDIO transport
