+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.0] - 2025-07-28
+
+### Added
+
+- Added support for STDIO transport through a Symfony Command : `mcp:stdio`
+- Added a new CommandTestCase for handling tests of the Symfony command
+
+### Updated
+
+- Updated the test to be more reusable, by creating a trait that contains DataProviders and assertion logic
+
 ## [1.3.0] - 2025-07-27
 
 ### Updated
