+++ src/EventListener/ExceptionListener.php
@@ -7,8 +7,10 @@ namespace Ecourty\McpServerBundle\EventListener;
 use Ecourty\McpServerBundle\Enum\McpErrorCode;
 use Ecourty\McpServerBundle\Exception\MethodHandlerNotFoundException;
 use Ecourty\McpServerBundle\Exception\RequestHandlingException;
+use Ecourty\McpServerBundle\Exception\ToolNotFoundException;
 use Ecourty\McpServerBundle\Service\ResponseFactory;
 use Symfony\Component\EventDispatcher\Attribute\AsEventListener;
+use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\RequestStack;
 use Symfony\Component\HttpKernel\Event\ExceptionEvent;
 use Symfony\Component\HttpKernel\Exception\UnprocessableEntityHttpException;
@@ -38,13 +40,28 @@ class ExceptionListener
 
         $response = match (true) {
             $exception instanceof UnprocessableEntityHttpException => $this->responseFactory->error($jsonRpcRequestId, McpErrorCode::PARSE_ERROR),
-            $exception instanceof MethodHandlerNotFoundException => $this->responseFactory->error($jsonRpcRequestId, McpErrorCode::TOOL_NOT_FOUND),
-            $exception instanceof RequestHandlingException => $this->responseFactory->error($jsonRpcRequestId, McpErrorCode::INTERNAL_ERROR),
+            $exception instanceof MethodHandlerNotFoundException => $this->responseFactory->error($jsonRpcRequestId, McpErrorCode::METHOD_NOT_FOUND),
+            $exception instanceof RequestHandlingException => $this->handleRequestHandlingException($exception, $jsonRpcRequestId),
             default => null,
         };
 
         if ($response !== null) {
             $event->setResponse($response);
         }
+    }
+
+    private function handleRequestHandlingException(
+        RequestHandlingException $exception,
+        string|int|null $jsonRpcRequestId = null,
+    ): JsonResponse {
+        $previous = $exception->getPrevious();
+        if ($previous instanceof ToolNotFoundException === true) {
+            return $this->responseFactory->error(
+                id: $jsonRpcRequestId,
+                errorCode: McpErrorCode::TOOL_NOT_FOUND,
+            );
+        }
+
+        return $this->responseFactory->error($jsonRpcRequestId, McpErrorCode::INTERNAL_ERROR);
     }
 }
