+++ src/Exception/ToolNotFoundException.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Ecourty\McpServerBundle\Exception;
+
+class ToolNotFoundException extends \Exception
+{
+}
